Change implementation of collection
Previously, a mutable set was used to contain the children of each
directory path. This was causing problems because these children are
mutable, and mutating an element in a hash set can create duplicates in
the set and make the element impossible to retrieve.

This commit replaces the hash set with a custom implementation of a
mutable set which fixes these problems. This implementation uses the
observer pattern to allow the contents of the set to notify it when they
change so that the set can rebuild itself. This imposes a performance
penalty whenever an item in the set changes state, but fixes any
potential problems that could be caused by the contents of the set
changing.

A set was chosen over a list because there is no need for the items in
the collection to be ordered, the collection must not contain duplicates
and checking membership is faster.