Replace methods with a property setter
This commit adds the functionality of some methods that modify a
property into the property setter. The reason why these were previously
separate is because other members of the class need to be able to modify
the property directly without invoking the methods. Now, an internal
backing property is used to allow that access.