Rename static factory method
This commit renames the static factory methods for creating paths to use
the `invoke` operator method. The purpose of this change is to allow the
client to create new instances as if they were calling a constructor.
Previously, they had to call a static method of the class to create new
objects.