Add custom collection implementation
This commit removes some methods for modifying a collection and moves
their functionality into a new implementation of that collection.
Previously, there could be confusion over whether the collection should
be modified directly or through the separate methods. Now, there is only
one way for the client to modify it. An internal backing property is
used to allow other members of the class to modify the collection
directly.