Add mutable collection implementation
Previously, there was a property for directory path objects that
returned a read-only set representing all the descendants of the
directory. This commit replaces it with an implementation of a mutable
set that allows for modifying the tree of paths. Adding paths to the set
inserts them into the proper location in the tree, and removing items
from the set removes them from the tree. The purpose of this change is
to provide a simpler interface for modifying the tree.