Replace operator method with regular one
Previously, a binary operator method was used for creating a copy of a
path with a given ancestor. This was problematic because the method
returned a copy of the path on the right side of the expression, and
binary operators typically return the object on the left side. This was
confusing behavior. The method that replaces the operator method
switches the receiver and the argument.