Make class immutable
Previously, a class for comparing directories used getters for computing
all of its properties. This commit replaces it with an immutable data
class that is created by a separate method.

This makes it possible to compare instances of the class. It also
prevents expensive filesystem operations from being executed on every
property access.