Add interfaces for new path type
The classes that were previously used for representing directory trees
had a complex type hierarchy that made them difficult to use and
required unsafe casts in some places. This complex type hierarchy also
made adding new file types difficult.

This commit introduces a new interface for representing directory trees.
This system uses the same interface for both files and directories, which
makes it much easier to use and requires fewer type checks and casts. It
also makes it much easier to add new file types. Where the previous
system assumed that anything that wasn't a directory was a file, this
system allows for paths with an unknown file type.