Prevent unnecessary copying
Previously, the only way to modify certain properties of path node
objects was to create a copy. In some methods, this resulted in
recursively copying entire trees of nodes. This commit makes those
properties mutable and adds setters to ensure that each node is always a
child of its parent.

By creating a copy of the node, it was previously possible to change
the node's file name. This caused unwanted behavior, and is no longer
possible.