Fix issue creating deep copies
Previously, creating a copy of a path node would deeply copy its
children but not its ancestors. This broke the contract that every path
node should be a child of its parent. You would end up with two nodes
with the same parent, and nodes can't have multiple children with the
same file name. This commit fixes the issue by making it so that copying
a node copies its children and ancestors deeply.