Remove feature
This commit removes the ability for path deltas to be applied to
multiple directories. The approach that was taken was problematic
because it wasn't flexible in how it handled filesystem actions
involving multiple directories (copy, move, etc.). It didn't work at all
when the source and target files were on different filesystems.