Simplify behavior of path deltas
Previously, path deltas allowed the user to apply a set of changes to
multiple directories by using relative paths. This was problematic for a
number of reasons. It doesn't work when the relative paths are
associated with different filesystems, as paths from different
filesystems can't be resolved against each other. It was also confusing
in some cases as to whether relative paths were being resolved against a
provided path or against the default directory of the filesystem.

This commit changes the behavior so that relative paths aren't treated
specially. They are always resolved against the default directory of the
filesystem. The purpose of this change is to simplify the usage, remove
edge cases and reduce the number of exceptions that can be thrown.