Convert interface to sealed class
Previously, an interface was used to provide a read-only view of a
mutable class. The problem with this was that there was nothing
preventing users from implementing the interface themselves, which means
that adding methods to the interface would be a backwards-incompatible
change. This commit replaces it with a sealed class which allows for
methods to be added in the future.