Add support for copying/moving between filesystems
The problem with recursively copying/moving files between filesystems is
that Java NIO provides no way to resolve paths that are associated with
different filesystems. This commit allows the user to provide a function
that converts paths between filesystems. While there's no way to
reliably convert paths between any two filesystems, this change gives
the user the ability to implement the functionality themselves for
whatever filesystems are important to them.